### 2025-01-24 #7709 @kyleniemeyer
Hi @oonjnertzern, thanks for your submission. Unfortunately, we have low editor availability right now, so I have to put your submission on our waitlist until someone frees up.

In the meantime, a few questions / comments:
- In the paper, please add an explicit `# References` section header at the end.
- I notice in your installation instructions in the README that you direct a conda-based environment, but then install your package using pip. Unless you specifically need conda (e.g., for a non-Python dependency), it would be simpler for folks to just install all dependencies with pip (and use venv to set up a virtual environment, for example).

  You could still include conda as alternate path for folks who use conda, but it doesn't seem necessary here (vs. a pure built-in Python installation route with venv & pip).
- Your documentation is missing some instructions on running the test suite, as well as any examples - could you add those?
