+++ docs/source/tutorials.rst
@@ -1,0 +1,8 @@
+Utilities
+==========
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Tutorials:
+
+   notebooks/Example 1
