+++ docs/requirements.txt
@@ -46,6 +46,8 @@ sphinxcontrib-serializinghtml==2.0.0
 tomli==2.0.1
 urllib3==2.2.3
 wheel==0.37.1
+ipykernel
+nbsphinx
 sphinx
 sphinx_rtd_theme
 myst-parser
