### 2025-02-11 7fbe739
Update conf.py
replaced nbsphinx with myst_nb
