+++ requirements_docs.txt
@@ -49,3 +49,4 @@ wheel==0.37.1
 sphinx
 sphinx_rtd_theme
 myst-parser
+nbsphinx
