### 2025-02-11 532aa25
moved readthedocs requirements file
