### 2025-02-11 b46daf3
added ipykernel to sphinx requirements
