+++ docs/requirements_docs.txt
@@ -50,3 +50,4 @@ sphinx
 sphinx_rtd_theme
 myst-parser
 nbsphinx
+ipykernel
