### 2025-02-11 4dc4e41
minor superficial edits to tutorial and install.rst
