### 2025-02-11 258caa0
edited install.rst and README to have same instructions
