+++ docs/source/index.rst
@@ -8,8 +8,6 @@ TEMPO documentation
 
 Welcome to the TEMPO documentation page! 
 
-Introduction
-==================
 
 TEMPO (Time-dependent Evolution of Multiple Pulse Operations) offers accessible and efficient simulations of pulse sequences in Python, using the suite of master equation solvers available in the Quantum Toolbox in Python (QuTiP). 
 It enables straightforward definition of pulse sequence structures, including any underlying time-dependent Hamiltonians and pulse timing information, and faster simulations of pulse sequence dynamics (compared to naive implementations using QuTiP) while remaining compatible with the existing collection of QuTiP subpackages. Given the ubiquitous use of pulse sequences throughout quantum information/computing sciences, magnetic resonance studies, and quantum metrology, this work has immediate relevance to a wide array of research applications.
@@ -21,9 +19,6 @@ Using the Docs
 The documentation page includes all primary modules of the library under the tab **TEMPO**. This includes the classes *Pulse*, *Pulse_recipe*, *Pulse_sequence*, *Hamiltonian*, and *Evolver*. Classes for exception handling and other miscellaneous operations are under **Utilities**. Please click on any of these modules for more details.
 You may also use the available search feature in the top left of the page, or the bottom of this one. 
 
-TEMPO Library 
-==================
-
 .. toctree::
    :maxdepth: 2
    :caption: Contents:
@@ -33,7 +28,7 @@ TEMPO Library
    tempo
    misc
 
-Indices and tables
+Index
 ==================
 
 * :ref:`modindex`
