### 2025-02-13 6940503
edits to code blocks in install.rst
