### 2025-02-13 e8dd4b4
edited README to match install.rst
