### 2025-02-13 #7709 @oonjnertzern
Hi @kyleniemeyer,

Thank you for making your acquaintance, and for getting things started for us.

Your suggestion for installation instructions using venv were especially helpful. I believe I have implemented all your suggestions, with changes to both the README and the documentation pages.

Please let us know if there is anything else we can help with, while we wait for editors to free up.

JJ.
