### 2025-02-17 b05e006
removed ascii art on README
