+++ README.md
@@ -1,3 +1,10 @@
+<!-- Badges -->
+<p align="center">
+  <a href="https://github.com/hakalasonja/tempo/actions/workflows/test.yml">
+    <img src="https://github.com/hakalasonja/tempo/actions/workflows/test.yml/badge.svg" alt="Build Status">
+  </a>
+</p>
+
 
 # Summary
 
