### 2025-04-01 #7709 @phibeck
Hi @oonjnertzern thanks for your patience. I'll be looking for reviewers next. When you have a moment it would be helpful if you could identify a few potential reviewers from this list https://reviewers.joss.theoj.org/reviewers (without tagging them with an @), thanks.
Additionally, please add an OSI-approved license to the repository, see https://joss.readthedocs.io/en/latest/review_criteria.html#software-license.
