### 2025-05-23 #8260 @editorialbot
[REVIEW]: TEMPO: A Python Package for Time Evolution of Pulse Sequences in QuTiP
**Submitting author:** @oonjnertzern (<a href="http://orcid.org/0000-0002-5716-5653">Jner Tzern Oon</a>)
**Repository:** https://github.com/hakalasonja/TEMPO
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
