### 2025-06-04 hakalasonja/TEMPO#2 @DanPuzzuoli
JOSS review: Citing existing software
Hi!

I'm just starting to look over the repo and the paper as part of the JOSS review. My first comment is that I think you should cite existing software that has pulse-sequence simulation capabilities (i.e. packages in which "pulses" can be defined, assembled into a "sequence", and then used in a dynamical simulation). Most relevant are the other packages that use QuTiP, e.g. [seQuencing](https://sequencing.readthedocs.io/en/latest/), but packages like [Qiskit Dynamics](https://joss.theoj.org/papers/10.21105/joss.05853) (shameless plug) also has these capabilities. It's of course difficult to create a comprehensive list, but the citations in the Qiskit Dynamics paper contains all the related packages I was aware of at the time of writing, so they are worth looking at and considering citing.

This will help to motivate the importance of what the package is meant to achieve, as well as provide more context when describing the different capabilities offered by TEMPO.
