### 2025-06-05 hakalasonja/TEMPO#4 @DanPuzzuoli
JOSS Review: Installation
I initially just tried following the installation process within a python 3.13 virtual environment (I always just default to the latest and try), but the installation failed. I think it was due to `qutip`, which is bound to `qutip < 5.0` in `requirements.txt` (and I guess doesn't work for python 3.13). I then tried again with a python 3.12 env and the process worked. It might be worth specifying somewhere (maybe just in the installation guide) which versions of python are supported.

Also, a question for @phibeck : the package is not actually installable as a python package. Is this a problem for JOSS standards? The documentation (e.g. in the tutorials) shows that to use the package the root folder should be added to the searchable directory list and then imported through relative imports. This should be documented in the installation guide.
