### 2025-06-06 hakalasonja/TEMPO#4 @phibeck
@DanPuzzuoli Thank you for raising this point in the review of https://github.com/openjournals/joss-reviews/issues/8260. I would say unless there's a specific reason why not, following the guidelines [here](https://joss.readthedocs.io/en/latest/review_criteria.html#installation-instructions) the package "should be `pip install`able".

@oonjnertzern could you take care of this and update the docs accordingly?
