### 2025-06-06 #8260 @phibeck
> [@phibeck](https://github.com/phibeck) I've added [another issue](https://github.com/hakalasonja/TEMPO/issues/4) about installation. Can you take a look and comment? (I ask you a question in it.)

Thanks @DanPuzzuoli - see my response in the issue. @oonjnertzern could you take a look? Thanks
