### 2025-08-06 d23227e
Extra section in paper for related pulse sequence software
