+++ paper/paper.bib
@@ -96,3 +96,82 @@
   month = mar 
 }
 
+@article{Alexander2020,
+  title = {Qiskit pulse: programming quantum computers through the cloud with pulses},
+  volume = {5},
+  ISSN = {2058-9565},
+  url = {http://dx.doi.org/10.1088/2058-9565/aba404},
+  DOI = {10.1088/2058-9565/aba404},
+  number = {4},
+  journal = {Quantum Science and Technology},
+  publisher = {IOP Publishing},
+  author = {Alexander,  Thomas and Kanazawa,  Naoki and Egger,  Daniel J and Capelluto,  Lauren and Wood,  Christopher J and Javadi-Abhari,  Ali and C McKay,  David},
+  year = {2020},
+  month = aug,
+  pages = {044006}
+}
+
+@article{Puzzuoli2023,
+  title = {Qiskit Dynamics: A Python package for simulating the
+time dynamics of quantum systems},
+  volume = {8},
+  ISSN = {2475-9066},
+  url = {http://dx.doi.org/10.21105/joss.05853},
+  DOI = {10.21105/joss.05853},
+  number = {90},
+  journal = {Journal of Open Source Software},
+  publisher = {The Open Journal},
+  author = {Puzzuoli,  Daniel and Wood,  Christopher J. and Egger,  Daniel J. and Rosand,  Benjamin and Ueda,  Kento},
+  year = {2023},
+  month = oct,
+  pages = {5853}
+}
+
+@misc{Sequencing,
+  doi = {10.5281/ZENODO.4515634},
+  url = {https://zenodo.org/record/4515634},
+  author = {Horn,  Logan Bishop-Van},
+  title = {sequencing-dev/sequencing: v1.2.0},
+  publisher = {Zenodo},
+  year = {2022},
+  copyright = {Open Access}
+}
+
+@article{Candoli2023,
+  title = {PULSEE: A software for the quantum simulation of an extensive set of magnetic resonance observables},
+  volume = {284},
+  ISSN = {0010-4655},
+  url = {http://dx.doi.org/10.1016/j.cpc.2022.108598},
+  DOI = {10.1016/j.cpc.2022.108598},
+  journal = {Computer Physics Communications},
+  publisher = {Elsevier BV},
+  author = {Candoli,  Davide and Nikolov,  Ilija K. and Brito,  Lucas Z. and Carr,  Stephen and Sanna,  Samuele and Mitrović,  Vesna F.},
+  year = {2023},
+  month = mar,
+  pages = {108598}
+}
+
+@article{Kramer2018,
+  title={QuantumOptics. jl: A Julia framework for simulating open quantum systems},
+  author={Kr{\"a}mer, Sebastian and Plankensteiner, David and Ostermann, Laurin and Ritsch, Helmut},
+  journal={Computer Physics Communications},
+  volume={227},
+  pages={109--116},
+  year={2018},
+  publisher={Elsevier}
+}
+
+@article{Hogben2011,
+  title = {Spinach – A software library for simulation of spin dynamics in large spin systems},
+  volume = {208},
+  ISSN = {1090-7807},
+  url = {http://dx.doi.org/10.1016/j.jmr.2010.11.008},
+  DOI = {10.1016/j.jmr.2010.11.008},
+  number = {2},
+  journal = {Journal of Magnetic Resonance},
+  publisher = {Elsevier BV},
+  author = {Hogben,  H.J. and Krzystyniak,  M. and Charnock,  G.T.P. and Hore,  P.J. and Kuprov,  Ilya},
+  year = {2011},
+  month = feb,
+  pages = {179–194}
+}
