### 2025-08-06 hakalasonja/TEMPO#2 @oonjnertzern
Thank you for your helpful comments @DanPuzzuoli ! I have added a section in the paper titled "Related Pulse Sequence Simulation Software" with some citations to this effect. Will be getting to your other comments next.
