### 2025-08-06 8c89d77
minor typo in citation reference
