### 2025-08-07 #8260 @oonjnertzern
Thanks for the reminder again @phibeck!
Sorry for the delays.
Yesterday I responded to thread [#2](https://github.com/hakalasonja/TEMPO/issues/2) by Dan Puzzuoli with some changes to the paper.
Am getting to the other comments in the coming days, thank you for your patience so far!
