### 2025-08-07 ad7c351
redundant requirements removed, small changes to pyproject and init files
