+++ tempo/__init__.py
@@ -1,1 +1,2 @@
-
+__version__ = "1.0.0"
+__doc__ = "A Python Package for Time Evolution of Pulse Sequences in QuTiP"
