+++ docs/requirements.txt
@@ -1,53 +1,0 @@
-alabaster==1.0.0
-attrs==24.2.0
-babel==2.16.0
-Brotli==1.1.0
-certifi==2024.8.30
-charset-normalizer==3.4.0
-dill==0.3.9
-docutils==0.21.2
-exceptiongroup==1.2.2
-idna==3.10
-imagesize==1.4.1
-iniconfig==1.1.1
-Jinja2==3.1.4
-markdown-it-py==3.0.0
-MarkupSafe==3.0.2
-mdit-py-plugins==0.4.2
-mdurl==0.1.2
-multiprocess==0.70.17
-myst-parser==4.0.0
-numpy==1.24.1
-packaging==23.2
-pip==23.2.1
-platformdirs==4.3.6
-pluggy==1.0.0
-pooch==1.8.2
-py==1.11.0
-Pygments==2.18.0
-pyparsing==3.0.9
-PySocks==1.7.1
-pytest==7.1.2
-PyYAML==6.0.2
-qutip==4.7.0
-requests==2.32.3
-scipy==1.10.0
-setuptools==63.4.1
-snowballstemmer==2.2.0
-Sphinx==8.1.3
-sphinx-rtd-theme==3.0.2
-sphinxcontrib-applehelp==2.0.0
-sphinxcontrib-devhelp==2.0.0
-sphinxcontrib-htmlhelp==2.1.0
-sphinxcontrib-jquery==4.1
-sphinxcontrib-jsmath==1.0.1
-sphinxcontrib-qthelp==2.0.0
-sphinxcontrib-serializinghtml==2.0.0
-tomli==2.0.1
-urllib3==2.2.3
-wheel==0.37.1
-ipykernel
-nbsphinx
-sphinx
-sphinx_rtd_theme
-myst-parser
