### 2025-08-20 9db2f5c
Changes for pip install and compatibility with qutip>5
