+++ requirements.txt
@@ -1,3 +1,3 @@
 numpy
-qutip<5.0
+qutip >= 4.7.0
 multiprocess
