### 2025-08-20 bd49edf
Made tests compatible with Qutip>5 changes
