### 2025-08-20 b81b7f7
tolerance edits and string formatting for hamiltonian.py
