### 2025-08-20 ae093f4
changes to evolver regular mehtod, and tests.py to use serial
