### 2025-08-21 e067333
restricted qutip to >5.0.0 and python to <3.13
