+++ requirements.txt
@@ -1,3 +1,3 @@
 numpy
-qutip >= 4.7.0
+qutip >= 5.0.0
 multiprocess
