### 2025-08-21 16bde59
changes to install.rst to reflect pip installation
