### 2025-08-21 hakalasonja/TEMPO#4 @oonjnertzern
Hi @DanPuzzuoli and @phibeck , thank you for these helpful comments! I believe the package is now pip-installable. I've updated the README and the readthedocs tutorials accordingly, but please let me know if I've missed something.

Also @DanPuzzuoli :
I took some time to update the code – it is now compatible with qutip>5.0.
Those changes should be reflected in the `requirements.txt` and the `pyproject.toml` files in the base directory. I've also listed python<3.13 as a requirement.
Please let me know if you run into any trouble!
