### 2025-09-05 #8260 @phibeck
Hi @oonjnertzern thanks for the update regarding https://github.com/hakalasonja/TEMPO/issues/4.

@DanPuzzuoli could you give this a check and see whether things are working as expected? Thanks!

@kareem1925 and @kaelynj please let me know if you have any questions along the way or if there's anything else holding up your reviews, thank you!
