### 2025-10-15 hakalasonja/TEMPO#4 @DanPuzzuoli
Just did the install process and started walking through one of the tutorials and it's all working. I'll close this issue and check the installation-related item off of my review. Thanks @oonjnertzern !
