### 2025-10-15 hakalasonja/TEMPO#2 @DanPuzzuoli
The updates look good to me, thanks!
